kafka-mqtt-connector/requirements.txt
confluent_kafka
paho-mqtt
